Lodha Codename Evergreen Vikhroli Mumbai
Welcome to Lodha Codename Evergreen, an exclusive development in a residential segment that brings to reality the most ideal and looked up to the normal of living, and holds the actual meaning of a 'peaceful lifestyle'. Positioned in one of the most beneficial and prominent locations of this city, at Vikhroli Mumbai, Lodha Codename Vikhroli Mumbai is an attractive compilation of exclusive 2 BHK Premium Residences that offer space and loneliness to you and your people, right in the heart of the city. 

For More Details :
Visit - https://lodha.developerprojects.in/residential/mumbai/lodha-codename-evergreen-vikhroli/